+++ Backend.Core/Backend.Core.csproj
@@ -7,6 +7,7 @@
 	</PropertyGroup>
 
 	<ItemGroup>
+		<PackageReference Include="Microsoft.AspNetCore.WebUtilities" Version="10.0.6" />
 		<PackageReference Include="Microsoft.Extensions.Options" Version="10.0.1" />
 		<PackageReference Include="NetVips" Version="3.0.0" />
 		<PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
