+++ Backend.EmailIngestion/Backend.EmailIngestion.csproj
@@ -10,6 +10,7 @@
 	<ItemGroup>
 		<PackageReference Include="MailKit" Version="4.7.0" />
 		<PackageReference Include="Microsoft.Extensions.Configuration" Version="10.0.1" />
+		<PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="10.0.2" />
 	</ItemGroup>
 
 	<ItemGroup>
