+++ Backend.WebApi/Cqrs/EmailProviderHandler.cs
@@ -1,5 +1,5 @@
+using Backend.EmailIngestion.Services;
 using Backend.WebApi.Cqrs.Infrastructure;
-using Backend.WebApi.Services;
 
 
 namespace Backend.WebApi.Cqrs;
