+++ Backend.WebApi/Cqrs/EmailQueryHandler.cs
@@ -1,26 +1,72 @@
 using Backend.Core.Authentication;
-using Backend.Core.Infrastructure;
+using Backend.EmailIngestion;
 using Backend.EmailIngestion.Services;
+using Backend.WebApi.Cqrs.Infrastructure;
+using Backend.WebApi.Services;
 using MailKit.Net.Imap;
 using MailKit.Security;
 
-namespace Backend.EmailIngestion;
+namespace Backend.WebApi.Cqrs;
 
-[RegisterService]
-public class ImapClientFactory
+
+[RequireAllowedTenantId]
+public class ListFolders : IQuery<ListFolders, IEnumerable<string>>
+{
+	public required Guid ExternalAccountId { get; set; }
+}
+
+
+[RequireAllowedTenantId]
+public class GetEmailContents : IQuery<GetEmailContents, GetEmailContents.EmailContents>
+{
+	public required Guid ExternalAccountId { get; set; }
+	public required string Folder { get; set; }
+	public required uint MessageId { get; set; }
+
+	public record EmailContents(string? Body, string? HtmlBody);
+}
+public class EmailQueryHandler :
+	IAsyncQueryHandler<ListFolders, IEnumerable<string>>,
+	IAsyncQueryHandler<GetEmailContents, GetEmailContents.EmailContents>
 {
 	private readonly ExternalAccountService _externalAccountService;
 	private readonly EmailProviderService _emailProviderService;
 	private readonly AccessTokenHelper _accessTokenHelper;
 
-	public ImapClientFactory(ExternalAccountService externalAccountService,
-		EmailProviderService emailProviderService,
-		AccessTokenHelper accessTokenHelper)
+	public EmailQueryHandler(ExternalAccountService externalAccountService,
+						  EmailProviderService emailProviderService,
+						  AccessTokenHelper accessTokenHelper)
 	{
 		_externalAccountService = externalAccountService;
 		_emailProviderService = emailProviderService;
 		_accessTokenHelper = accessTokenHelper;
 	}
+
+
+	public async Task<IEnumerable<string>> Handle(ListFolders query)
+	{
+		var imapClient = await GetImapClient(query.ExternalAccountId);
+
+		return await imapClient.GetAvailableFolders();
+	}
+
+
+	public async Task<GetEmailContents.EmailContents> Handle(GetEmailContents query)
+	{
+		var imapClient = await GetImapClient(query.ExternalAccountId);
+
+		var emailContents = await imapClient.GetEmailContents(query.Folder, query.MessageId);
+		if (emailContents == null)
+		{
+			throw new Exception($"Email with MessageId {query.MessageId} not found in folder {query.Folder}");
+		}
+
+		return new GetEmailContents.EmailContents(
+			emailContents.PlainText,
+			emailContents.Html
+		);
+	}
+
 
 	private async Task<IImapClient> GetImapClient(Guid externalAccountId)
 	{
@@ -58,26 +104,4 @@ public class ImapClientFactory
 		}
 	}
 
-
-	public static async Task<IImapClient> ConnectAsync(IAuthContext auth, string emailAddress, EmailProviderSettings.EmailProvider emailProvider)
-	{
-		var imapClient = new ImapClient();
-		await imapClient.ConnectAsync(emailProvider.ImapHost, emailProvider.ImapPort, SecureSocketOptions.SslOnConnect);
-
-		switch (auth)
-		{
-			case OAuthContext oauth:
-				await imapClient.AuthenticateAsync(new SaslMechanismOAuth2(emailAddress, oauth.AccessToken));
-				break;
-
-			case BasicAuthContext basicAuth:
-				await imapClient.AuthenticateAsync(basicAuth.Username, basicAuth.Password);
-				break;
-
-			default:
-				throw new ArgumentException("Unsupported auth context type.");
-		}
-
-		return imapClient;
-	}
 }
